remove compiler warnings

+++ uint/uint64_pack.c
@@ -4,5 +4,5 @@
 
 void uint64_pack(char *out,uint64 in) {
   uint32_pack(out,in&0xffffffff);
-  uint32_pack(out+4,in>>32);
+  uint32_pack(out+4,(uint32)(in>>32));
 }
